# src/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Inventory API",
    "version": "1.0.0",
    "description": "REST API for products and stock levels. Writes require an X-API-Key header when the server has API_KEY set."
  },
  "components": {
    "securitySchemes": {
      "ApiKey": { "type": "apiKey", "in": "header", "name": "X-API-Key" }
    },
    "parameters": {
      "id": { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } },
      "ifMatch": { "name": "If-Match", "in": "header", "required": false, "description": "ETag from GET /products/{id}; 412 if stale. Optional.", "schema": { "type": "string" } },
      "ifNoneMatch": { "name": "If-None-Match", "in": "header", "required": false, "description": "Returns 304 if it matches the current ETag.", "schema": { "type": "string" } }
    },
    "schemas": {
      "ProductInput": {
        "type": "object",
        "required": ["name", "sku", "price", "quantity"],
        "properties": {
          "name": { "type": "string", "minLength": 1, "maxLength": 100 },
          "sku": { "type": "string", "minLength": 1, "maxLength": 40 },
          "price": { "type": "number", "minimum": 0 },
          "quantity": { "type": "integer", "minimum": 0 },
          "reorderLevel": { "type": "integer", "minimum": 0 },
          "category": { "type": "string", "minLength": 1, "maxLength": 50 }
        }
      },
      "Product": {
        "allOf": [
          { "$ref": "#/components/schemas/ProductInput" },
          { "type": "object", "required": ["id"], "properties": { "id": { "type": "integer" } } }
        ]
      },
      "ProductPage": {
        "type": "object",
        "properties": {
          "items": { "type": "array", "items": { "$ref": "#/components/schemas/Product" } },
          "total": { "type": "integer" },
          "limit": { "type": "integer" },
          "offset": { "type": "integer" }
        }
      },
      "MovementInput": {
        "type": "object",
        "required": ["type", "quantity", "reason"],
        "properties": {
          "type": { "type": "string", "enum": ["in", "out", "adjust"] },
          "quantity": { "type": "integer", "minimum": 0 },
          "reason": { "type": "string", "minLength": 1, "maxLength": 200 },
          "reference": { "type": "string", "maxLength": 100 }
        }
      },
      "Movement": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "productId": { "type": "integer" },
          "type": { "type": "string", "enum": ["in", "out", "adjust"] },
          "delta": { "type": "integer" },
          "balance": { "type": "integer" },
          "reason": { "type": "string" },
          "reference": { "type": "string" },
          "at": { "type": "integer", "description": "Epoch milliseconds" }
        }
      },
      "Error": {
        "type": "object",
        "description": "RFC 9457 problem details",
        "required": ["type", "title", "status", "detail", "error"],
        "properties": {
          "type": { "type": "string", "example": "about:blank" },
          "title": { "type": "string" },
          "status": { "type": "integer" },
          "detail": { "type": "string" },
          "error": { "description": "Legacy field: message or Zod flattened error", "oneOf": [{ "type": "string" }, { "type": "object" }] },
          "errors": { "type": "object", "description": "Zod field errors (400 validation only)", "additionalProperties": { "type": "array", "items": { "type": "string" } } }
        }
      }
    },
    "responses": {
      "BadRequest": { "description": "Validation error", "content": { "application/problem+json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
      "NotFound": { "description": "Not found", "content": { "application/problem+json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
      "Unauthorized": { "description": "Missing or wrong API key", "content": { "application/problem+json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
    }
  },
  "paths": {
...
    "/products": {
      "get": {
        "summary": "List products",
        "parameters": [
          { "name": "limit", "in": "query", "schema": { "type": "integer", "minimum": 1, "maximum": 100, "default": 20 } },
          { "name": "offset", "in": "query", "schema": { "type": "integer", "minimum": 0, "default": 0 } },
          { "name": "sort", "in": "query", "schema": { "type": "string", "enum": ["price", "name"] } },
          { "name": "order", "in": "query", "schema": { "type": "string", "enum": ["asc", "desc"], "default": "asc" } },
          { "name": "q", "in": "query", "description": "Search name or SKU", "schema": { "type": "string", "maxLength": 100 } },
          { "name": "category", "in": "query", "description": "Only products in this category (case-insensitive)", "schema": { "type": "string", "maxLength": 50 } },
          { "name": "lowStock", "in": "query", "description": "Only products with quantity <= N", "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": { "description": "Page of products", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ProductPage" } } } },
          "400": { "$ref": "#/components/responses/BadRequest" }
        }
      },
      "post": {
        "summary": "Create a product",
        "description": "Optionally send an Idempotency-Key header. A repeat with the same key and body replays the stored response (with Idempotent-Replayed: true); same key with a different body returns 422. Keys are kept in memory for 24 h.",
        "security": [{ "ApiKey": [] }],
        "parameters": [
          { "name": "Idempotency-Key", "in": "header", "required": false, "schema": { "type": "string" } }
        ],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ProductInput" } } } },
        "responses": {
          "201": { "description": "Created", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Product" } } } },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "409": { "description": "SKU already exists", "content": { "application/problem+json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
          "422": { "description": "Idempotency-Key reused with a different body", "content": { "application/problem+json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
...
    "/products/import": {
      "post": {
        "summary": "Import products from CSV",
        "description": "Header must include name, sku, price, quantity; reorderLevel and category are optional.",
        "security": [{ "ApiKey": [] }],
        "requestBody": { "required": true, "content": { "text/csv": { "schema": { "type": "string" } } } },
        "responses": {
          "200": {
            "description": "Per-row import result",
            "content": { "application/json": { "schema": { "type": "object", "properties": { "imported": { "type": "integer" }, "failed": { "type": "integer" }, "rows": { "type": "array", "items": { "type": "object", "properties": { "row": { "type": "integer" }, "ok": { "type": "boolean" }, "errors": { "type": "array", "items": { "type": "string" } } } } } } } } }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
    "/products/{id}": {
      "parameters": [{ "$ref": "#/components/parameters/id" }],
      "get": {
        "summary": "Get a product",
        "parameters": [{ "$ref": "#/components/parameters/ifNoneMatch" }],
        "responses": {
          "304": { "description": "Not modified (If-None-Match matched)" },
          "200": { "description": "Product", "headers": { "ETag": { "description": "Version of the product; changes on any stock or detail change", "schema": { "type": "string" } } }, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Product" } } } },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      },
      "patch": {
        "summary": "Edit product details (name, price, reorderLevel, category); stock is unchanged",
        "security": [{ "ApiKey": [] }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "minProperties": 1, "additionalProperties": false, "properties": { "name": { "type": "string", "minLength": 1, "maxLength": 100 }, "price": { "type": "number", "minimum": 0 }, "reorderLevel": { "type": "integer", "minimum": 0 }, "category": { "type": "string", "minLength": 1, "maxLength": 50 } } } } } },
        "responses": {
          "200": { "description": "Updated product", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Product" } } } },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      },
      "delete": {
        "summary": "Delete a product",
        "security": [{ "ApiKey": [] }],
        "responses": {
          "204": { "description": "Deleted" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/products/{id}/stock": {
      "parameters": [{ "$ref": "#/components/parameters/id" }],
      "patch": {
        "summary": "Adjust stock by a delta",
        "security": [{ "ApiKey": [] }],
        "parameters": [{ "$ref": "#/components/parameters/ifMatch" }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "required": ["delta"], "properties": { "delta": { "type": "integer" } } } } } },
        "responses": {
          "200": { "description": "Updated product", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Product" } } } },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "412": { "description": "If-Match did not match the current ETag", "content": { "application/problem+json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
          "422": { "description": "Insufficient stock", "content": { "application/problem+json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
    "/stock/batch": {
      "post": {
        "summary": "Atomically apply up to 100 stock adjustments",
        "description": "All-or-nothing: if any item is invalid, unknown or would make stock negative, nothing is applied. Items for the same product are applied in order and checked cumulatively. One movement is recorded per non-zero item.",
        "security": [{ "ApiKey": [] }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "required": ["items"], "properties": { "items": { "type": "array", "minItems": 1, "maxItems": 100, "items": { "type": "object", "required": ["id", "delta"], "properties": { "id": { "type": "integer" }, "delta": { "type": "integer" }, "reason": { "type": "string", "maxLength": 200 } } } } } } } } },
        "responses": {
          "200": { "description": "Updated products, one per distinct id in first-seen order", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Product" } } } } },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "413": { "description": "More than 100 items", "content": { "application/problem+json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
          "422": { "description": "One or more items failed; `errors` lists {index, id, reason}", "content": { "application/problem+json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
    "/products/{id}/movements": {
      "parameters": [{ "$ref": "#/components/parameters/id" }],
      "get": {
        "summary": "Stock movement history, newest first",
        "responses": {
          "200": { "description": "Movements", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Movement" } } } } },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      },
      "post": {
        "summary": "Record a stock movement",
        "parameters": [{ "$ref": "#/components/parameters/ifMatch" }],
        "description": "type 'adjust' sets stock to the absolute counted quantity.",
        "security": [{ "ApiKey": [] }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/MovementInput" } } } },
        "responses": {
          "201": { "description": "Created", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Movement" } } } },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "412": { "description": "If-Match did not match the current ETag", "content": { "application/problem+json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    }
  }
}